+++ header.php
@@ -1,8 +1,8 @@
 <?php
 /**
  * The header template.
- * @package highwind light
- * @since 1.0
+ * @package highwind_light
+ * @since 1.0.0
  */
 ?>
 <?php if ( ! defined( 'ABSPATH' ) ) exit; // Exit if accessed directly ?><?php highwind_html_before(); ?><!doctype html><!--[if lt IE 7 ]> <html <?php language_attributes(); ?> class="no-js ie6"> <![endif]-->
@@ -43,7 +43,7 @@
 	<?php highwind_header_before(); ?>
 
 	<?php
-	if( is_home() ) {
+	if( is_home() || is_front_page() ) {
 		$header_image = header_image();
 		$header_color = '';
 	}
@@ -54,13 +54,21 @@
 		else{
 			$post_id 			= get_the_ID(); // Get the Post ID
 		}
-		$post_thumbnail_id 	= get_post_thumbnail_id( $post_id ); // Get the Attachment ID from post
-		$post_image 		= wp_get_attachment_image_src( $post_thumbnail_id, 'full' );
-		$header_image 		= apply_filters( 'highwind_header_featured_image_size', $post_image[0] );
-		$header_color 		= get_post_meta( $post_id, 'highwind-header-color', true );
+		$post_thumbnail_id 		= get_post_thumbnail_id( $post_id ); // Get the Attachment ID from post
+		$post_image 			= wp_get_attachment_image_src( $post_thumbnail_id, 'full' );
+		$header_image 			= apply_filters( 'highwind_header_featured_image_size', $post_image[0] );
+		$header_color 			= get_post_meta( $post_id, 'highwind-header-color', true );
+		$header_bg_position_x 	= get_post_meta( $post_id, 'highwind-header-bg-position-x', true );
+		$header_bg_position_y 	= get_post_meta( $post_id, 'highwind-header-bg-position-y', true );
+		$header_bg_repeat 		= get_post_meta( $post_id, 'highwind-header-bg-repeat', true );
+
+		// If no setting has been found, set default.
+		if( empty( $header_bg_position_x ) ) { $header_bg_position_x = 'center'; }
+		if( empty( $header_bg_position_y ) ) { $header_bg_position_y = 'center'; }
+		if( empty( $header_bg_repeat ) ) { $header_bg_repeat = 'no-repeat'; }
 	}
 	?>
-	<header class="header content-wrapper" role="banner" style="background-color:<?php echo $header_color; ?>; background-image:url(<?php echo $header_image; ?>);">
+	<header class="header content-wrapper" role="banner" style="background-color:<?php echo $header_color; ?>; background-image:url(<?php echo $header_image; ?>); background-position-x:<?php echo $header_bg_position_x; ?>; background-position-y:<?php echo $header_bg_position_y; ?>; background-repeat:<?php echo $header_bg_repeat; ?>;">
 
 		<?php highwind_header(); ?>
 
@@ -69,3 +77,16 @@
 	<div class="content-wrapper">
 
 	<?php highwind_header_after(); ?>
+
+	<?php
+	/**
+	 * WordPress SEO by Yoast - Breadcrumbs
+	 *
+	 * @since 1.1.2
+	 */
+	if ( get_theme_mod( 'highwind_light_breadcrumbs' ) == 'yes' ) {
+		if ( function_exists('yoast_breadcrumb') ) {
+			yoast_breadcrumb('<p id="breadcrumbs">','</p>');
+		}
+	}
+	?>