+++ projects/templates/single-project/pagination.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * Single project pagination
+ *
+ * @author 		Sebastien Dumont
+ * @package 	Highwind Light/Projects/Templates
+ * @version 	1.1.2
+ */
+
+/*
+ * The single project pagination has been removed for this theme as one is already in place.
+ */
+
+if ( ! defined( 'ABSPATH' ) ) exit; // Exit if accessed directly
+?>