+++ projects/templates/single-project/title.php
@@ -1,0 +1,16 @@
+<?php
+/**
+ * Single Project title
+ *
+ * @author 		Sebastien Dumont
+ * @package 	Highwind Light/Projects/Templates
+ * @version 	1.1.2
+ */
+
+/*
+ * The single project title has been removed for this theme as it is already in the header.
+ */
+
+if ( ! defined( 'ABSPATH' ) ) exit; // Exit if accessed directly
+
+?>