+++ changelog.md
@@ -2,8 +2,11 @@
 
 ## 1.1.2 - currently developing
 * ADDED - The ability to hide the title in the header per post/page to show only the featured image. Useful for artwork or photos.
-* ADDED - Support for Projects by WooThemes
+* ADDED - Support for Projects by WooThemes.
 * ADDED - 404 page.
+* ADDED - Breadcrumb support for WordPress SEO by Yoast. Can also be switched on or off in the theme customizer.
+* ADDED - The ability to alter the featured image position and repeat values for the header.
+* ADDED - The ability to add a featured image for categories and taxonomies. - Currently working on.
 * CORRECTED - Featured Image is shown in the content only on archive pages.
 * CORRECTED - Header title link error if on a 404 page.
 * CORRECTED - Header colour options if page is not a page or post.